First attempt at documenting surveys of interest
This adds two tables to the README:
- One table that contains the list of surveys discussed, reviewed, or prepared
in the WebDX Community Group. That table is meant to provide a quick overview
of the group's work with links to individual surveys. It will grow over time.
- One table that sets rough milestones for recurring surveys.

It would be useful to collect additional information about each survey,
including links to issues that they gave birth to, e.g.,  in the Interop
project, CSS WG, or wherever data was used. Not done here!